backend/requirements.txt
fastapi
uvicorn
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv
pydantic
requests
openai>=1.12.0